Minor header size tweaks, make backup of css
I've finally taken the time to learn just a BIT of Sass and install extensions for it. I've yet to test the auto-compiler, and I expect it to overwrite main.css when I do, so I've made a backup of my edits to main.css for me to reference as I migrate over to working on the sass.